split push/deploy integration tests across files and add tier 2 cases
